Adds stacked output Y getter
The D3 stackLayout object's x, y & out methods are rebound to the
stackedBar. Currently, the stackedBar is using the `yValue` accessor the
access the value of the datum. This works for 'zero' based stacks
as the value of the datum is equal to the stack layout's outputted
'y' value.

However, when using offset algorithms such as 'expand', we need
to plot the value outputted from the stackLayout, rather than the Y value
from the datum. The method `stackY` was added which defines an accessor
for the stackLayout's output Y value, much like the existing and renamed
`stackY0`.