Decorate stacked series with 'stacked' class
Rebinding the decorate directly to the multi series isn't possible
as the stacked bar needs to decorate the series itself.

The stacked series provides a decorate() function to the
underlying multi series that adds the class, and then forwards
the decoration on to the decorate() implementation supplied
by the user.