Refactors stacked series to use multi series
The stacked.stack function is basically replicating
the functionality of the multi series. All a stacked
series is, is multiple bar/line/area series with different base
lines.

The stacked series provides a decorate() function to the
underlying multi series that adds the class, and then forwards
the decoration on to the decorate() implementation supplied
by the user. Rebinding the decorate directly to the multi series
isn't possible as the stacked bar needs to decorate the series
with the 'stacked' class.